Backend Set-up Complete.
The back-end of this project is complete. The idea is to keep it simple for now.
I have used the Swagger UI for accessing the back-end functionalities, without the need for a front-end as of now.
I intend to keep working on it and add a frontend.

It has a keyword search feature that allows the users to search through the job posts based on a keyword.
A wildcard search implementation in on the todo.

Peace out. :D